Fix ST-GCN Phase 2 float64/float32 dtype mismatch
DTUDataset yields float64 EEG tensors (raw .mat arrays are double-precision);
the graph-conv Chebyshev basis buffer is float32, so torch.matmul in
GraphConvKW.forward raised "expected scalar type Float but found Double".
Cast eeg to float32 at both train/eval call sites in the training notebook.

The earlier CUDA "no kernel image available" error (P100/sm_60) is gone in
this run's log -- Kaggle assigned a T4 this time -- so this was the only
remaining blocker.

+++ notebooks/kaggle_train_stgcn_gcn_only.py
@@ -219,7 +219,9 @@ def run_fold(subject_id, fold, tr_split, te_split, keep_curve: bool):
         model.train()
         tr_loss_sum, tr_correct, tr_n = 0.0, 0, 0
         for eeg, _audio, y in train_loader:
-            eeg, y = eeg.to(DEVICE), y.to(DEVICE).long()
+            # DTUDataset yields float64 (the raw .mat EEG arrays are double-precision);
+            # the graph-conv basis buffer is float32, so this cast is required.
+            eeg, y = eeg.to(DEVICE).float(), y.to(DEVICE).long()
             opt.zero_grad()
             logits = model(eeg)
             loss = loss_fn(logits, y)
@@ -233,7 +235,7 @@ def run_fold(subject_id, fold, tr_split, te_split, keep_curve: bool):
         te_loss_sum, te_correct, te_n = 0.0, 0, 0
         with torch.no_grad():
             for eeg, _audio, y in test_loader:
-                eeg, y = eeg.to(DEVICE), y.to(DEVICE).long()
+                eeg, y = eeg.to(DEVICE).float(), y.to(DEVICE).long()
                 logits = model(eeg)
                 loss = loss_fn(logits, y)
                 te_loss_sum += loss.item() * y.size(0)
